EnhancedJSONEncoder is centralized
