+++ encoders.py
@@ -1,4 +1,5 @@
 import json
+from dataclasses import asdict, is_dataclass
 from datetime import datetime
 
 
@@ -7,3 +8,10 @@ class DateTimeEncoder(json.JSONEncoder):
         if isinstance(obj, datetime):
             return obj.isoformat()
         return super().default(obj)
+
+
+class EnhancedJSONEncoder(json.JSONEncoder):
+    def default(self, o):
+        if is_dataclass(o):
+            return asdict(o)
+        return super().default(o)
