+++ geo_fence_operations/urls.py
@@ -1,21 +1,6 @@
-"""flight_blender URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/3.1/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.urls import path
-from . import views as geo_fence_views
 
+from . import views as geo_fence_views
 
 urlpatterns = [
     path("set_geo_fence", geo_fence_views.set_geo_fence, name="set_geo_fence"),
